Backend/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.10.8
django-allauth==0.33.0
django-cors-headers==2.1.0
django-rest-auth==0.9.1
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
facebook-sdk==2.0.0
idna==2.6
oauthlib==2.0.3
PyJWT==1.5.3
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.22
mysqlclient==1.3.13
